+++ src/main/java/net/lopymine/ip/family/generation/TextureGenerationManager.java
@@ -1,6 +1,7 @@
 package net.lopymine.ip.family.generation;
 
 import com.mojang.blaze3d.platform.NativeImage;
+import java.io.IOException;
 import java.util.*;
 import java.util.Map.Entry;
 import java.util.concurrent.*;
@@ -15,6 +16,7 @@ import net.lopymine.ip.utils.iac.RenderedItemImage;
 import net.minecraft.client.Minecraft;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.resources.Identifier;
+import net.minecraft.util.Util;
 import net.minecraft.world.item.*;
 import org.jetbrains.annotations.Nullable;