+++ src/main/java/net/lopymine/ip/mixin/GuiMixin.java
@@ -1,0 +1,41 @@
+package net.lopymine.ip.mixin;
+
+import net.lopymine.ip.config.InventoryParticlesConfig;
+import net.lopymine.ip.config.sub.InventoryParticlesMainConfig;
+import net.lopymine.ip.renderer.InventoryParticlesRenderer;
+import net.minecraft.client.gui.Gui;
+import net.minecraft.client.gui.screens.Screen;
+import org.spongepowered.asm.mixin.*;
+import org.spongepowered.asm.mixin.injection.*;
+import org.spongepowered.asm.mixin.injection.At.Shift;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(Gui.class)
+public class GuiMixin {
+
+	//? if >=26.2 {
+	@Inject(at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/screens/Screen;removed()V"), method = "setScreen")
+	private void clearRendererWhenRemovedScreen(Screen screen, CallbackInfo ci) {
+		InventoryParticlesMainConfig config = InventoryParticlesConfig.getInstance().getMainConfig();
+		if (!config.isModEnabled()) {
+			return;
+		}
+		InventoryParticlesRenderer.getInstance().clear();
+	}
+
+	@Inject(at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/screens/Screen;init(II)V", shift = Shift.AFTER), method = "setScreen")
+	private void initRendererWhenInitScreen(Screen screen, CallbackInfo ci) {
+		this.initRenderer();
+	}
+
+	@Unique
+	private void initRenderer() {
+		InventoryParticlesMainConfig config = InventoryParticlesConfig.getInstance().getMainConfig();
+		if (!config.isModEnabled()) {
+			return;
+		}
+		InventoryParticlesRenderer.getInstance().init();
+	}
+	//?}
+
+}