+++ src/main/java/net/lopymine/ip/mixin/MinecraftClientMixin.java
@@ -5,6 +5,7 @@ import net.lopymine.ip.config.sub.InventoryParticlesMainConfig;
 import net.lopymine.ip.family.atlas.manager.FamilyParticlesAtlasManager;
 import net.lopymine.ip.renderer.InventoryParticlesRenderer;
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.Gui;
 import net.minecraft.client.gui.screens.Screen;
 import com.mojang.blaze3d.platform.Window;
 import org.jetbrains.annotations.Nullable;
@@ -16,14 +17,29 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(Minecraft.class)
 public class MinecraftClientMixin {
 
-	@Shadow @Nullable public Screen screen;
-
-	@Shadow @Final private Window window;
+	@Shadow
+	@Final
+	private Window window;
 
 	@Inject(at = @At("HEAD"), method = "close")
-	private void inject(CallbackInfo ci) {
+	private void closeFamilyParticleAtlases(CallbackInfo ci) {
 		FamilyParticlesAtlasManager.closeAll();
 	}
+
+	//? if >=26.2 {
+	@Shadow @Final public Gui gui;
+
+	@Inject(
+			at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/screens/Screen;resize(II)V"),
+			method = "resizeGui"
+	)
+	private void updateParticlesPositions(CallbackInfo ci) {
+		this.updateParticlesPositions();
+	}
+	//?} else <=26.1 {
+	/*@Shadow
+	@Nullable
+	public Screen screen;
 
 	@Inject(at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/screens/Screen;removed()V"), method = "setScreen")
 	private void clearRendererWhenRemovedScreen(Screen screen, CallbackInfo ci) {
@@ -45,14 +61,14 @@ public class MinecraftClientMixin {
 			//? if >=26.1 {
 			method = "resizeGui"
 			//?} else {
-			/*method = "resizeDisplay"
-			*///?}
+			/^method = "resizeDisplay"
+			^///?}
 	)
 	private void updateParticlesPositions(CallbackInfo ci) {
 		this.updateParticlesPositions();
 	}
 	//?} else {
-	/*@Inject(at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/screens/Screen;init(Lnet/minecraft/client/Minecraft;II)V", shift = Shift.AFTER), method = "setScreen")
+	/^@Inject(at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/screens/Screen;init(Lnet/minecraft/client/Minecraft;II)V", shift = Shift.AFTER), method = "setScreen")
 	private void initRendererWhenInitScreen(Screen screen, CallbackInfo ci) {
 		this.initRenderer();
 	}
@@ -61,7 +77,7 @@ public class MinecraftClientMixin {
 	private void updateParticlesPositions(CallbackInfo ci) {
 		this.updateParticlesPositions();
 	}
-	*///?}
+	^///?}
 
 	@Unique
 	private void initRenderer() {
@@ -72,9 +88,15 @@ public class MinecraftClientMixin {
 		InventoryParticlesRenderer.getInstance().init();
 	}
 
+	*///?}
+
 	@Unique
 	private void updateParticlesPositions() {
-		Screen screen = this.screen;
+		//? if >=26.2 {
+		Screen screen = this.gui.screen();
+		//?} else {
+		/*Screen screen = this.screen;
+		 *///?}
 		if (screen == null) {
 			return;
 		}